requirements.txt
pyyaml
streamlit
